backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# AI & LLM
openai==1.3.5
httpx==0.25.1

# Authentication
firebase-admin==6.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment & Config
python-dotenv==1.0.0
pydantic-settings==2.0.3
email-validator==2.2.0

# File Processing
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
python-slugify==8.0.1
python-magic==0.4.27
aiofiles==23.2.1

# Logging & Monitoring
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.10.1
flake8==6.1.0
isort==5.12.0
